+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -14,14 +14,15 @@ namespace Thumber\Test\TestCase\TestSuite;
 
 use Cake\Core\Configure;
 use Thumber\TestSuite\TestCase;
-use Thumber\ThumbTrait;
+use Thumber\ThumbsPathTrait;
+use Thumber\Utility\ThumbManager;
 
 /**
  * TestCaseTest class
  */
 class TestCaseTest extends TestCase
 {
-    use ThumbTrait;
+    use ThumbsPathTrait;
 
     /**
      * Test for `assertImageFileEquals()` method
@@ -29,12 +30,12 @@ class TestCaseTest extends TestCase
      */
     public function testAssertImageFileEquals()
     {
-        $original = Configure::read(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg';
+        $original = Configure::readOrFail(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg';
         $copy = tempnam(TMP, $original);
 
         copy($original, $copy);
 
-        $this->assertImageFileEquals(Configure::read(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg', $copy);
+        $this->assertImageFileEquals(Configure::readOrFail(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg', $copy);
         $this->assertImageFileEquals('resize_w200_h200.jpg', $copy);
     }
 
@@ -44,7 +45,7 @@ class TestCaseTest extends TestCase
      */
     public function testAssertThumbPath()
     {
-        foreach ($this->getSupportedFormats() as $extension) {
+        foreach (ThumbManager::$supportedFormats as $extension) {
             $this->assertThumbPath($this->getPath() . DS . md5(time()) . '_' . md5(time()) . '.' . $extension);
         }
     }