+++ tests/TestCase/Utility/ThumbCreatorSaveTest.php
@@ -12,8 +12,8 @@
  */
 namespace Thumber\Test\TestCase\Utility;
 
+use Cake\Core\Configure;
 use Thumber\TestSuite\TestCase;
-use Thumber\ThumbTrait;
 use Thumber\Utility\ThumbCreator;
 
 /**
@@ -21,8 +21,6 @@ use Thumber\Utility\ThumbCreator;
  */
 class ThumbCreatorSaveTest extends TestCase
 {
-    use ThumbTrait;
-
     /**
      * Test for `save()` method
      * @test
@@ -37,7 +35,7 @@ class ThumbCreatorSaveTest extends TestCase
         ];
 
         //Adds some extensions only for the `imagick` driver
-        if ($this->getDriver() == 'imagick') {
+        if (Configure::readOrFail(THUMBER . '.driver') == 'imagick') {
             $extensions += [
                 'bmp' => 'image/x-ms-bmp',
                 'ico' => 'image/x-icon',
@@ -74,7 +72,7 @@ class ThumbCreatorSaveTest extends TestCase
      */
     public function testSaveFromInvalidFileGd()
     {
-        $this->skipIf($this->getDriver() != 'gd');
+        $this->skipIfDriverIs('imagick');
 
         (new ThumbCreator(APP . 'config' . DS . 'routes.php'))->resize(200)->save();
     }
@@ -89,7 +87,7 @@ class ThumbCreatorSaveTest extends TestCase
      */
     public function testSaveFromInvalidFileImagick()
     {
-        $this->skipIf($this->getDriver() != 'imagick');
+        $this->skipIfDriverIs('gd');
 
         (new ThumbCreator(APP . 'config' . DS . 'routes.php'))->resize(200)->save();
     }
@@ -198,7 +196,7 @@ class ThumbCreatorSaveTest extends TestCase
         $file = (new ThumbCreator('400x400.png'))->resize(200)->save(['format' => 'jpeg']);
         $this->assertFileExtension('jpg', $file);
 
-        $this->skipIf($this->getDriver() === 'gd');
+        $this->skipIfDriverIs('gd');
 
         $file = (new ThumbCreator('400x400.png'))->resize(200)->save(['format' => 'tif']);
         $this->assertFileExtension('tiff', $file, PATHINFO_EXTENSION);
@@ -218,7 +216,7 @@ class ThumbCreatorSaveTest extends TestCase
     /**
      * Test for `save()` method, using the `target` option with a no existing
      *  directory target
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessageRegExp /^The directory `[\w\/:\\]+` is not writeable$/
      * @test
      */
@@ -230,7 +228,7 @@ class ThumbCreatorSaveTest extends TestCase
 
     /**
      * Test for `save()` method, without a valid method called before
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessage No valid method called before the `save` method
      */
     public function testSaveWithoutCallbacks()